requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
dj-static==0.0.6
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
static3==0.7.0
typing_extensions==4.12.2
tzdata==2024.1
